dynamic analysis for context

+++ tests/run_test.py
@@ -4,7 +4,9 @@ import subprocess
 result_file = open('helper-progtype.txt','w')
 result_file.write(f"{'function_name':<40}{'program_type':<40}{'compile error':<20}{'compile warning':<20}{'allowed'}\n")
 
-program_types = ['cgroup/dev', 'cgroup/skb','cgroup_skb/egress', 'cgroup_skb/ingress', 'cgroup/getsockopt', 'cgroup/setsockopt', 'cgroup/bind4',
+program_types = ['cgroup/connect_unix','freplace/print']
+'''
+        'cgroup/dev', 'cgroup/skb','cgroup_skb/egress', 'cgroup_skb/ingress', 'cgroup/getsockopt', 'cgroup/setsockopt', 'cgroup/bind4',
         'cgroup/bind6', 'cgroup/connect6', 'cgroup/getpeername6', 'cgroup/getsockname6', 'cgroup/recvmsg4', 'cgroup/sendmsg4', 'cgroup/recvmsg6',
         'cgroup/sendmsg6', 'cgroup/connect_unix', 'cgroup/sendmsg_unix', 'cgroup/recvmsg_unix', 'cgroup/getpeername_unix', 'cgroup/getsockname_unix', 
         'cgroup/post_bind4', 'cgroup/post_bind6', 'cgroup/sock_create', 'cgroup/sock', 'cgroup/sock_release', 'cgroup/sysctl',  
@@ -17,6 +19,7 @@ program_types = ['cgroup/dev', 'cgroup/skb','cgroup_skb/egress', 'cgroup_skb/ing
         'fexit/do_nanosleep', 'fexit.s/bpf_fentry_test1', 'iter/tcp', 'iter.s/cgroup', 'tp_btf/task_newtask', 'xdp.frags/cpumap', 'xdp/cpumap',
         'xdp.frags/devmap', 'xdp/devmap', 'xdp', 'xdp.frags']
 '''
+'''
 other_program_types = ['freplace',  'struct_ops'] 
 '''
 
@@ -26,26 +29,6 @@ for program_type in program_types:
     prog_type_file = open("create_tests/prog_type.txt", "w")
     prog_type_file.write(program_type)
     prog_type_file.close()
-    '''
-    create_test_program = open("create_tests/create_test.py", "r")
-    edited_lines = []
-    while True:
-        line = create_test_program.readline()
-        if not line:
-            break
-        if "attach" in line:
-            print("okay")
-            edited_lines.append("attach_type ="+program_type)
-        else:
-            edited_lines.append(line)
-    create_test_program.close()
-
-    for line in edited_lines:
-        print(line) 
-    create_test_program = open("create_tests/create_test.py", "w")
-    create_test_program.writelines(edited_lines)
-    create_test_program.close()
-    '''    
     
     for i in range(216):
         os.system('python3 create_tests/create_test.py')
@@ -67,10 +50,19 @@ for program_type in program_types:
         result_file.write(f"{program_type:<40}")
 
         #os.system('sudo make test 2>make_result.txt')
-        make_command = ['make', 'test']
-        make_file = open("make_result.txt", 'w')
-        subprocess.call(make_command, stderr=make_file, shell=False)
-        make_file.close()
+        if program_type == "freplace/print":
+            make_command = ['clang', '-O2', '-g', '-target', 'bpf', '-c', 'test.bpf.c', '-o', 'test.bpf.o']
+            make_file = open("make_result.txt", 'w')
+            subprocess.call(make_command, stderr = make_file, shell=False)
+            make_command = ['gcc', '-o', 'freplace', 'freplace.c', '-lbpf']
+            subprocess.call(make_command, stderr = make_file, shell=False)
+            make_file.close()
+        
+        else:
+            make_command = ['make', 'test']
+            make_file = open("make_result.txt", 'w')
+            subprocess.call(make_command, stderr=make_file, shell=False)
+            make_file.close()
 
         make_file = open('make_result.txt', 'r')
         make_result = make_file.read().replace('\n', '')
@@ -87,7 +79,10 @@ for program_type in program_types:
 
         #os.system('sudo ./test &>run_result.txt')
         run_file = open("run_result.txt", 'w')
-        subprocess.call('./test', stderr = run_file, shell=False)
+        if program_type=="freplace/print":
+            subprocess.call('./freplace', stderr = run_file, shell=False)
+        else:
+            subprocess.call('./test', stderr = run_file, shell=False)
         run_file.close()
 
         run_file = open("run_result.txt", "r")
